requirements.txt
# This file is here because many Platforms as a Service look for
#	requirements.txt in the root directory of a project.
-r requirements/production.txt
django-pipeline==1.4.6
jsmin==2.1.1
django-redis==3.8.3
elasticsearch==1.4.0
django-haystack==2.3.1
